+++ test/anvil/auth/oidc_test.exs
@@ -1,5 +1,5 @@
 defmodule Anvil.Auth.OIDCTest do
-  use ExUnit.Case, async: true
+  use Supertester.ExUnitFoundation, isolation: :full_isolation
 
   alias Anvil.Auth.OIDC