+++ test/anvil/export/csv_test.exs
@@ -8,6 +8,7 @@ defmodule Anvil.Export.CSVTest do
   setup do
     # Start Ecto Sandbox
     :ok = Ecto.Adapters.SQL.Sandbox.checkout(Anvil.Repo)
+    Ecto.Adapters.SQL.Sandbox.mode(Anvil.Repo, {:shared, self()})
 
     # Create test data
     {:ok, labeler1} =