+++ test/anvil/forge_bridge/sample_dto_test.exs
@@ -1,5 +1,5 @@
 defmodule Anvil.ForgeBridge.SampleDTOTest do
-  use ExUnit.Case, async: true
+  use Supertester.ExUnitFoundation, isolation: :full_isolation
 
   alias Anvil.ForgeBridge.SampleDTO