+++ test/test_helper.exs
@@ -1,7 +1,9 @@
-ExUnit.start()
+ExUnit.start(max_cases: 1)
 
 # Configure Supertester
 Application.put_env(:supertester, :isolation, :full_isolation)
 
-# Start Ecto Sandbox for test isolation
-Ecto.Adapters.SQL.Sandbox.mode(Anvil.Repo, :manual)
+# Start Ecto Sandbox for test isolation when repo is enabled
+if Application.get_env(:anvil, :start_repo, true) do
+  Ecto.Adapters.SQL.Sandbox.mode(Anvil.Repo, :manual)
+end